% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example.R
\name{readroper_example}
\alias{readroper_example}
\title{Get path to readroper example}
\usage{
readroper_example(path = NULL)
}
\arguments{
\item{path}{Name of file. If `NULL`, the example files will be listed.}
}
\description{
readroper comes bundled with a number of sample files in its `inst/extdata`
directory. This function make them easy to access
}
\examples{
readroper_example()
readroper_example('processData.R')
}
